Revert "Merge pull request #334 from navikt/support-local-test"
This reverts commit 2687726e036e5f12bd46ff26091bfb0094cfbaf5, reversing
changes made to ee7fcdfcfc3bd1aad80b971550bc617ce2a3c604.
