+++ api/src/main/resources/tilbakemeldingsmottak-api.yml
@@ -7,8 +7,6 @@ info:
 servers:
   - url: http://localhost:9069
     description: Lokal utvikling
-security:
-  - bearerAuth: [ ]
 paths:
   /rest/taskserviceklage/klassifiser:
     description: Saksbehandler klassifiserer serviceklage som er sendt inn.
@@ -197,7 +195,6 @@ paths:
       tags:
         - nais-contract
       operationId: isAlive
-      security: [ ]
       responses:
         "200":
           description: OK
@@ -211,7 +208,6 @@ paths:
       tags:
         - nais-contract
       operationId: selftest
-      security: [ ]
       responses:
         "200":
           description: OK
@@ -225,7 +221,6 @@ paths:
       tags:
         - nais-contract
       operationId: isReady
-      security: [ ]
       responses:
         "200":
           description: OK
@@ -252,11 +247,6 @@ paths:
           - health
 
 components:
-  securitySchemes:
-    bearerAuth:
-      type: http
-      scheme: bearer
-      bearerFormat: JWT
   schemas:
     ApplicationStatus:
       type: object
