+++ app/src/main/kotlin/no/nav/tilbakemeldingsmottak/config/ClaimChecker.kt
@@ -1,15 +1,12 @@
-package no.nav.tilbakemeldingsmottak.config.security
+package no.nav.tilbakemeldingsmottak.config
 
 import org.slf4j.LoggerFactory
-import org.springframework.beans.factory.annotation.Value
 import org.springframework.security.core.Authentication
 import org.springframework.security.oauth2.server.resource.authentication.JwtAuthenticationToken
 import org.springframework.stereotype.Component
 
 @Component("claimChecker")
-class ClaimChecker(
-    @Value("\${auth.issuers.azuread.issuer-uri}") private val azureadIssuer: String
-) {
+class ClaimChecker {
 
     private val log = LoggerFactory.getLogger(javaClass)
 
@@ -21,8 +18,8 @@ class ClaimChecker(
 
         val jwt = authentication.token
         val issuer = jwt.issuer?.toString() ?: return false
-        if (!issuer.equals(azureadIssuer, ignoreCase = true)) {
-            log.info("Avvist: issuer $issuer er ikke konfigurert AzureAD issuer")
+        if (!issuer.contains("login.microsoftonline.com", ignoreCase = true)) {
+            log.info("Avvist: issuer $issuer er ikke AzureAD")
             return false
         }
