+++ app/src/main/resources/application.yml
@@ -32,6 +32,3 @@ buffersize: 1
 logging:
   level:
     org.hibernate.orm.connections.pooling: WARN
-
-springdoc:
-  enable-data-rest: false
