--- app/src/test/resources/application-disableauthtest.yml
@@ -1,3 +1,0 @@
-app:
-  security:
-    disable-auth: true
