--- app/src/test/resources/application-localtest.yml
@@ -1,8 +1,0 @@
-spring:
-  config:
-    activate:
-      on-profile: localtest
-
-app:
-  security:
-    disable-auth: true
